Fix style

+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/program/DefaultProgramInstanceService.java
@@ -110,11 +110,33 @@ public class DefaultProgramInstanceService
 
     @Override
     @Transactional
+    public long addProgramInstance( ProgramInstance programInstance, User user )
+    {
+        programInstanceStore.save( programInstance, user );
+        return programInstance.getId();
+    }
+
+    @Override
+    @Transactional
     public void deleteProgramInstance( ProgramInstance programInstance )
     {
-        programInstance.setStatus( ProgramStatus.CANCELLED );
-        programInstanceStore.update( programInstance );
-        programInstanceStore.delete( programInstance );
+        deleteProgramInstance( programInstance, false );
+    }
+
+    @Override
+    @Transactional
+    public void deleteProgramInstance( ProgramInstance programInstance, boolean forceDelete )
+    {
+        if ( forceDelete )
+        {
+            programInstanceStore.delete( programInstance );
+        }
+        else
+        {
+            programInstance.setDeleted( true );
+            programInstance.setStatus( ProgramStatus.CANCELLED );
+            programInstanceStore.update( programInstance );
+        }
     }
 
     @Override
@@ -209,17 +231,14 @@ public class DefaultProgramInstanceService
             throw new IllegalQueryException( "Program does not exist: " + program );
         }
 
-        TrackedEntityType te =
-            trackedEntityType != null ? trackedEntityTypeService.getTrackedEntityType( trackedEntityType ) : null;
+        TrackedEntityType te = trackedEntityType != null ? trackedEntityTypeService.getTrackedEntityType( trackedEntityType ) : null;
 
         if ( trackedEntityType != null && te == null )
         {
             throw new IllegalQueryException( "Tracked entity does not exist: " + program );
         }
 
-        TrackedEntityInstance tei = trackedEntityInstance != null ?
-            trackedEntityInstanceService.getTrackedEntityInstance( trackedEntityInstance ) :
-            null;
+        TrackedEntityInstance tei = trackedEntityInstance != null ? trackedEntityInstanceService.getTrackedEntityInstance( trackedEntityInstance ) : null;
 
         if ( trackedEntityInstance != null && tei == null )
         {
@@ -324,33 +343,24 @@ public class DefaultProgramInstanceService
         {
             if ( !aclService.canDataRead( params.getUser(), params.getProgram() ) )
             {
-                throw new IllegalQueryException(
-                    "Current user is not authorized to read data from selected program:  " +
-                        params.getProgram().getUid() );
+                throw new IllegalQueryException( "Current user is not authorized to read data from selected program:  " + params.getProgram().getUid() );
             }
 
-            if ( params.getProgram().getTrackedEntityType() != null &&
-                !aclService.canDataRead( params.getUser(), params.getProgram().getTrackedEntityType() ) )
+            if ( params.getProgram().getTrackedEntityType() != null && !aclService.canDataRead( params.getUser(), params.getProgram().getTrackedEntityType() ) )
             {
-                throw new IllegalQueryException(
-                    "Current user is not authorized to read data from selected program's tracked entity type:  " +
-                        params.getProgram().getTrackedEntityType().getUid() );
+                throw new IllegalQueryException( "Current user is not authorized to read data from selected program's tracked entity type:  " + params.getProgram().getTrackedEntityType().getUid() );
             }
 
         }
 
-        if ( params.hasTrackedEntityType() &&
-            !aclService.canDataRead( params.getUser(), params.getTrackedEntityType() ) )
+        if ( params.hasTrackedEntityType() && !aclService.canDataRead( params.getUser(), params.getTrackedEntityType() ) )
         {
-            throw new IllegalQueryException(
-                "Current user is not authorized to read data from selected tracked entity type:  " +
-                    params.getTrackedEntityType().getUid() );
+            throw new IllegalQueryException( "Current user is not authorized to read data from selected tracked entity type:  " + params.getTrackedEntityType().getUid() );
         }
     }
 
     @Override
-    public void validate( ProgramInstanceQueryParams params )
-        throws IllegalQueryException
+    public void validate( ProgramInstanceQueryParams params ) throws IllegalQueryException
     {
         String violation = null;
 
@@ -361,14 +371,12 @@ public class DefaultProgramInstanceService
 
         User user = params.getUser();
 
-        if ( !params.hasOrganisationUnits() &&
-            !(params.isOrganisationUnitMode( ALL ) || params.isOrganisationUnitMode( ACCESSIBLE )) )
+        if ( !params.hasOrganisationUnits() && !(params.isOrganisationUnitMode( ALL ) || params.isOrganisationUnitMode( ACCESSIBLE )) )
         {
             violation = "At least one organisation unit must be specified";
         }
 
-        if ( params.isOrganisationUnitMode( ACCESSIBLE ) &&
-            (user == null || !user.hasDataViewOrganisationUnitWithFallback()) )
+        if ( params.isOrganisationUnitMode( ACCESSIBLE ) && (user == null || !user.hasDataViewOrganisationUnitWithFallback()) )
         {
             violation = "Current user must be associated with at least one organisation unit when selection mode is ACCESSIBLE";
         }
@@ -432,8 +440,7 @@ public class DefaultProgramInstanceService
 
     @Override
     @Transactional( readOnly = true )
-    public List<ProgramInstance> getProgramInstances( TrackedEntityInstance entityInstance, Program program,
-        ProgramStatus status )
+    public List<ProgramInstance> getProgramInstances( TrackedEntityInstance entityInstance, Program program, ProgramStatus status )
     {
         return programInstanceStore.get( entityInstance, program, status );
     }
@@ -441,14 +448,11 @@ public class DefaultProgramInstanceService
     @Override
     @Transactional
     public ProgramInstance prepareProgramInstance( TrackedEntityInstance trackedEntityInstance, Program program,
-        ProgramStatus programStatus, Date enrollmentDate, Date incidentDate, OrganisationUnit organisationUnit,
-        String uid )
+        ProgramStatus programStatus, Date enrollmentDate, Date incidentDate, OrganisationUnit organisationUnit, String uid )
     {
-        if ( program.getTrackedEntityType() != null &&
-            !program.getTrackedEntityType().equals( trackedEntityInstance.getTrackedEntityType() ) )
+        if ( program.getTrackedEntityType() != null && !program.getTrackedEntityType().equals( trackedEntityInstance.getTrackedEntityType() ) )
         {
-            throw new IllegalQueryException(
-                "Tracked entity instance must have same tracked entity as program: " + program.getUid() );
+            throw new IllegalQueryException( "Tracked entity instance must have same tracked entity as program: " + program.getUid() );
         }
 
         ProgramInstance programInstance = new ProgramInstance();
@@ -497,8 +501,7 @@ public class DefaultProgramInstanceService
         // Add program instance
         // ---------------------------------------------------------------------
 
-        ProgramInstance programInstance = prepareProgramInstance( trackedEntityInstance, program, ProgramStatus.ACTIVE,
-            enrollmentDate,
+        ProgramInstance programInstance = prepareProgramInstance( trackedEntityInstance, program, ProgramStatus.ACTIVE, enrollmentDate,
             incidentDate, organisationUnit, uid );
         addProgramInstance( programInstance );
 
@@ -507,6 +510,7 @@ public class DefaultProgramInstanceService
         // ---------------------------------------------------------------------
 
         trackerOwnershipAccessManager.assignOwnership( trackedEntityInstance, program, organisationUnit, true, true );
+
 
         // -----------------------------------------------------------------
         // Send enrollment notifications (if any)
@@ -562,8 +566,7 @@ public class DefaultProgramInstanceService
         // Send sms-message after program completion
         // ---------------------------------------------------------------------
 
-        eventPublisher
-            .publishEvent( new ProgramEnrollmentCompletionNotificationEvent( this, programInstance.getId() ) );
+        eventPublisher.publishEvent( new ProgramEnrollmentCompletionNotificationEvent( this, programInstance.getId() ) );
 
         eventPublisher.publishEvent( new EnrollmentEvaluationEvent( this, programInstance.getId() ) );
     }
@@ -615,8 +618,7 @@ public class DefaultProgramInstanceService
         {
             log.warn( "Program has another active enrollment going on. Not possible to incomplete" );
 
-            throw new IllegalQueryException(
-                "Program has another active enrollment going on. Not possible to incomplete" );
+            throw new IllegalQueryException( "Program has another active enrollment going on. Not possible to incomplete" );
         }
 
         // -----------------------------------------------------------------
